+++ backend/lambdas/dataProcess/csvCommon/db/model/Rataosoite.ts
