+++ backend/lambdas/dataProcess/handleCSVFileEvent/csvDataParser/__tests__/csvDataParser.test.ts
@@ -177,6 +177,7 @@ describe('handle ams csv file success', () => {
        amsCsvStream,
        {},
      );
+     console.log(result);
   }, 900000);
 });
